requirements.txt
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
certifi==2021.10.8
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==2.0.7
constantly==15.1.0
cryptography==41.0.2
daphne==4.0.0
Django==3.2.8
django-admin-index==2.0.2
django-crontab==0.7.1
django-i18next==0.1.3
django-ipware==4.0.0
django-js-asset==2.0.0
django-mptt==0.14.0
django-nine==0.2.5
django-ordered-model==3.7.4
djangorestframework==3.12.4
et-xmlfile==1.1.0
hyperlink==21.0.0
idna==3.2
importlib-metadata==6.7.0
incremental==22.10.0
msgpack==1.0.5
oauthlib==3.1.1
openpyxl==3.0.9
pathlib2==2.3.6
Pillow==8.4.0
psycopg2==2.9.1
py3dns==3.2.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==23.2.0
pyTelegramBotAPI==4.7.0
python-crontab==2.5.1
python-dateutil==2.8.2
pytz==2021.3
redis==4.6.0
requests==2.26.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
Twisted==22.10.0
txaio==23.1.1
typing-extensions==4.7.1
urllib3==1.26.7
uWSGI==2.0.20
validate-email==1.3
whatsapp-api-client-python==0.0.31
zipp==3.15.0
zope.interface==6.0
